We need to better align the usage of the filecatalog and daq db with their intended usage.  The filelists table in the daqdb is the primary authority for what files have been produced by the experiment and are available for event building.  The datasets table in the filecatlog is the primary authority for the availabity of derivative files which are available for downstream analysis.  The files table in the filecatalog keeps track of the location of files on the various filesystems.
So.  We need to modify the input queries in the configuration file such
that we can specify both the query and the target database.  This will
enable us to obtain a list of available files for each run from the
filelists (daqdb) table or the datasets (filecatalog) table.  Then, the
consistency check may be executed against the files (filecatalog) table
without needing to specify a user query in the configuration.

This commit simply confirms that the proposed alteration to the configuration
file can be properly parsed.